oracle-ai-database-gcp-gemini/python/requirements.txt
# Core Google Cloud & Vertex AI
google-cloud-aiplatform>=1.70.0
vertexai

# Google ADK and MCP
google-adk>=1.25.1
google-genai>=1.0.0
mcp>=0.9.0

# MCP Toolbox for Databases (recommended for production)
toolbox-core>=0.5.0

# LangChain for RAG
langchain>=0.3.7
langchain-core>=0.3.15
langchain-community>=0.3.5
langchain-text-splitters>=0.3.2
langchain-google-vertexai>=2.0.5
langchain-google-genai
langchain-huggingface

# Oracle Database
oracledb>=2.0.0
oci>=2.165.0

# FastAPI for RAG API server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pydantic>=2.9.2

# PDF processing
PyPDF2>=3.0.1

# Utilities
python-dotenv>=1.0.1
requests>=2.32.4
tqdm
streamlit
sentence-transformers
